pushing changes to mhd_class-refactor
