+++ lib/foods.js
@@ -1,5 +1,4 @@
 const shared = require('./shared.js')
-
 const $ = require('jQuery')
 const api = 'https://lit-caverns-20261.herokuapp.com'
