+++ lib/shared.js
@@ -1,5 +1,6 @@
 const $ = require('jQuery')
 const api = 'https://lit-caverns-20261.herokuapp.com'
+const requests = require('./requests')
 
 function foodRows(food) {
   let foodName = food.name
@@ -31,71 +32,14 @@ function focusListener(foodId) {
     $(this).children('[name]').each(function(index) {
       toPass.food[$(this).attr('name')] = $.trim($(this).text());
     })
-    updateFood(id, toPass)
+    requests.updateFood(id, toPass)
   });
 };
 
-function updateFood(id, toPass) {
-  $.ajax({
-    method: 'PATCH',
-    url: api + `/api/v1/foods/${id}`,
-    data: toPass,
-    success: function(data) {
-      console.log(`${data.name} successfully updated!`);
-    }
-  })
-}
 
 function deleteRow() {
     let foodId = $(this).parent().parent().data().id
-    deleteMeals(foodId);
+    requests.getMealsForItemDelete(foodId)
 };
-
-function deleteFoodMeals(meals, foodId) {
-    meals.forEach(function(meal, index) {
-       $.ajax({
-        method: 'DELETE',
-        url: api + `/api/v1/meals/${meal.id}/foods/${foodId}`,
-        success: function() {
-          if (index === (meals.length - 1)) {
-            deleteFood(foodId);
-          }
-        }
-      });
-    })
-};
-
-function deleteMeals(foodId) {
-   $.ajax({
-    method: 'GET',
-    url: api + '/api/v1/meals',
-    success: function(data) {
-      meals = findMealsWithFood(foodId, data);
-      if (meals.length) {
-        deleteFoodMeals(meals, foodId);
-      } else {
-        deleteFood(foodId);
-      }
-    }
-  })
-};
-
-function findMealsWithFood(foodId, meals) {
-  return meals.filter(function(meal) {
-    return meal.foods.some(function(food) {
-      return food.id === foodId;
-    })
-  })
-}
-
-function deleteFood(id) {
-  $.ajax({
-    method: 'DELETE',
-    url: api + `/api/v1/foods/${id}`,
-    success: function() {
-      $(`[data-id=${id}]`).remove();
-    }
-  })
-}
 
 module.exports = {foodRows, iconListener, focusListener, deleteRow}
